Add new files and make code improvements
