+++ PoliwhiRL/models/RainbowDQN/ReplayBuffer.py
@@ -1,5 +1,6 @@
+# -*- coding: utf-8 -*-
 import numpy as np
-import torch
+
 
 class PrioritizedReplayBuffer:
     def __init__(self, capacity, alpha=0.6):
@@ -10,12 +11,14 @@ class PrioritizedReplayBuffer:
         self.priorities = np.zeros((capacity,), dtype=np.float32)  # Store priorities
 
     def add(self, state, action, reward, next_state, done, error):
-        max_prio = self.priorities.max() if self.buffer else 1.0  # Max priority for new entry
+        max_prio = (
+            self.priorities.max() if self.buffer else 1.0
+        )  # Max priority for new entry
         if len(self.buffer) < self.capacity:
             self.buffer.append((state, action, reward, next_state, done))
         else:
             self.buffer[self.pos] = (state, action, reward, next_state, done)
-        
+
         self.priorities[self.pos] = max_prio if error is None else error
         self.pos = (self.pos + 1) % self.capacity
 
@@ -23,18 +26,18 @@ class PrioritizedReplayBuffer:
         if len(self.buffer) == self.capacity:
             prios = self.priorities
         else:
-            prios = self.priorities[:self.pos]
+            prios = self.priorities[: self.pos]
 
-        probs = prios ** self.alpha
+        probs = prios**self.alpha
         probs /= probs.sum()
-        
+
         indices = np.random.choice(len(self.buffer), batch_size, p=probs)
         samples = [self.buffer[idx] for idx in indices]
 
         total = len(self.buffer)
         weights = (total * probs[indices]) ** (-beta)
         weights /= weights.max()  # Normalize for stability
-        
+
         states, actions, rewards, next_states, dones = zip(*samples)
         return states, actions, rewards, next_states, dones, indices, weights
